+++ mantrap/solver/solver_intermediates/z_controls.py
@@ -19,14 +19,14 @@ class ZControlIntermediate(Solver, ABC):
     # Transformations #########################################################
     ###########################################################################
     def z_to_ego_trajectory(self, z: np.ndarray, return_leaf: bool = False) -> torch.Tensor:
-        ego_controls = torch.from_numpy(z).view(self.planning_horizon, 2)
+        ego_controls = torch.from_numpy(z).view(self.planning_horizon, 2).float()
         ego_controls.requires_grad = True
         ego_trajectory = self.env.ego.unroll_trajectory(controls=ego_controls, dt=self.env.dt)
         assert check_ego_trajectory(ego_trajectory, t_horizon=self.planning_horizon + 1, pos_and_vel_only=True)
         return ego_trajectory if not return_leaf else (ego_trajectory, ego_controls)
 
     def z_to_ego_controls(self, z: np.ndarray, return_leaf: bool = False) -> torch.Tensor:
-        ego_controls = torch.from_numpy(z).view(self.planning_horizon, 2)
+        ego_controls = torch.from_numpy(z).view(self.planning_horizon, 2).float()
         ego_controls.requires_grad = True
         assert check_ego_controls(ego_controls, t_horizon=self.planning_horizon)
         return ego_controls if not return_leaf else (ego_controls, ego_controls)
